stock_assistant/requirements.txt
langchain_core
langgraph
langchain_openai
python-dotenv
aiohttp
flask
flask[async]
flask-cors
langchain_community
mysql-connector-python
pinecone